+++ src/Application/Deployment/DeploymentStatusClient.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace EventFlow\Application\Deployment;
+
+interface DeploymentStatusClient
+{
+    public function get(string $url): DeploymentStatusResponse;
+}